Release 2.9.0: cut version, move GitHub links into About, refresh feature list
- Cut the unreleased changelog section as 2.9.0 and bump package.json to 2.9.0.
  Coach stays on the published npm package (@perchwerks/eye-in-the-sky@0.5.0) —
  BETA has been on it since the 2.7.0 cut (never flipped back to the git dep), so
  it's already in release config; no flip needed for the main push.
- Move the open-source GitHub repo links off the landing page and into the About
  dialog, under the "Free & Open Source" section, above the feature list. Shown on
  all platforms there (consistent with Credits), so the native landing stays lean.
- Refresh the About feature list with the headline additions from the last month
  (7 languages, native Android app + MyChron Wi-Fi download, phone-as-logger GPS
  timing, optional cloud sync, AI coach), translated across all locales.
