+++ src/locales/de/landing.json
@@ -106,7 +106,12 @@
       "Editor für eigene Strecken & Kurse mit Community-Einsendungen",
       "Lokale Wetterabfrage",
       "Dunkel- & Hell-Modus",
-      "PWA — installierbar & vollständig offline"
+      "PWA — installierbar & vollständig offline",
+      "In 7 Sprachen verfügbar",
+      "Native Android-App – WLAN-Download vom AiM MyChron",
+      "Smartphone als Logger – GPS-Rundenzeiten ohne Zusatzhardware",
+      "Optionale Cloud-Synchronisierung über deine Geräte (kostenlose Stufe)",
+      "KI-Fahrcoach (optionales Add-on)"
     ]
   },
   "supportedFiles": {
