+++ src/locales/en/landing.json
@@ -105,7 +105,12 @@
       "Custom track & course editor with community submissions",
       "Local weather lookup",
       "Dark & light mode",
-      "PWA — installable & fully offline"
+      "PWA — installable & fully offline",
+      "Available in 7 languages",
+      "Native Android app — Wi-Fi download from AiM MyChron",
+      "Phone-as-logger — GPS lap timing with no extra hardware",
+      "Optional cloud sync across your devices (free tier)",
+      "AI driving coach (optional add-on)"
     ]
   },
   "fileImport": {
