+++ src/locales/es/landing.json
@@ -106,7 +106,12 @@
       "Editor de pistas y circuitos personalizados con envíos comunitarios",
       "Consulta meteorológica local",
       "Modo oscuro y claro",
-      "PWA: instalable y totalmente sin conexión"
+      "PWA: instalable y totalmente sin conexión",
+      "Disponible en 7 idiomas",
+      "App nativa de Android: descarga por Wi-Fi desde AiM MyChron",
+      "El teléfono como registrador: cronometraje por GPS sin hardware extra",
+      "Sincronización en la nube opcional entre tus dispositivos (nivel gratuito)",
+      "Entrenador de pilotaje con IA (complemento opcional)"
     ]
   },
   "supportedFiles": {
