+++ src/locales/fr/landing.json
@@ -106,7 +106,12 @@
       "Éditeur de circuits et de parcours personnalisés avec soumissions communautaires",
       "Consultation de la météo locale",
       "Mode sombre et clair",
-      "PWA — installable et entièrement hors ligne"
+      "PWA — installable et entièrement hors ligne",
+      "Disponible en 7 langues",
+      "App Android native — téléchargement Wi-Fi depuis AiM MyChron",
+      "Téléphone comme enregistreur — chronométrage GPS sans matériel supplémentaire",
+      "Synchronisation cloud optionnelle entre vos appareils (offre gratuite)",
+      "Coach de pilotage IA (option payante)"
     ]
   },
   "supportedFiles": {
