+++ src/locales/it/landing.json
@@ -106,7 +106,12 @@
       "Editor di piste e circuiti personalizzati con invii della community",
       "Consultazione meteo locale",
       "Modalità scura e chiara",
-      "PWA — installabile e completamente offline"
+      "PWA — installabile e completamente offline",
+      "Disponibile in 7 lingue",
+      "App Android nativa — download Wi-Fi da AiM MyChron",
+      "Telefono come logger — cronometraggio GPS senza hardware aggiuntivo",
+      "Sincronizzazione cloud opzionale tra i tuoi dispositivi (piano gratuito)",
+      "Coach di guida con IA (componente aggiuntivo opzionale)"
     ]
   },
   "supportedFiles": {
