+++ src/locales/ja/landing.json
@@ -106,7 +106,12 @@
       "コミュニティ送信付きのカスタムトラック・コースエディター",
       "ローカル天気の取得",
       "ダーク＆ライトモード",
-      "PWA — インストール可能で完全オフライン"
+      "PWA — インストール可能で完全オフライン",
+      "7 言語に対応",
+      "ネイティブ Android アプリ — AiM MyChron から Wi-Fi でダウンロード",
+      "スマホをロガーに — 追加ハードウェア不要の GPS ラップタイム計測",
+      "デバイス間の任意のクラウド同期（無料プラン）",
+      "AI ドライビングコーチ（オプションのアドオン）"
     ]
   },
   "supportedFiles": {
