+++ src/locales/pt-BR/landing.json
@@ -106,7 +106,12 @@
       "Editor de pistas e circuitos personalizados com envios da comunidade",
       "Consulta de clima local",
       "Modo escuro e claro",
-      "PWA — instalável e totalmente offline"
+      "PWA — instalável e totalmente offline",
+      "Disponível em 7 idiomas",
+      "App nativo de Android — download por Wi-Fi do AiM MyChron",
+      "Celular como logger — cronometragem por GPS sem hardware extra",
+      "Sincronização na nuvem opcional entre seus dispositivos (plano gratuito)",
+      "Treinador de pilotagem com IA (complemento opcional)"
     ]
   },
   "supportedFiles": {
